rename: wishcard 컴포넌트 이름 templeStayCard로 변경
