+++ src/components/card/templeStayCard/InfoSection.css.ts
